@@ -6,9 +6,12 @@ import { MatInputModule } from '@angular/material/input';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
 import { MatMenuModule } from '@angular/material/menu';
-import { FormsModule } from '@angular/forms';
-import { MatSnackBar, matSnackBarAnimations, MatSnackBarModule } from '@angular/material/snack-bar';
+import { FormsModule } from "@angular/forms";
+import { MatSelectModule } from '@angular/material/select';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
 import { HttpClientModule } from '@angular/common/http';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatDividerModule } from '@angular/material/divider';
 
 
 import { AppRoutingModule } from './app-routing.module';
@@ -30,9 +33,13 @@ import { ProductFormComponent } from './product-form/product-form.component';
     MatInputModule,
     MatToolbarModule,
     MatButtonModule,
+    MatMenuModule,
     FormsModule,
     MatSnackBarModule,
-    HttpClientModule
+    HttpClientModule,
+    MatSelectModule,
+    MatDialogModule,
+    MatDividerModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
