@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
+
+import {MatButtonModule} from '@angular/material/button';
 
 interface Assistant {
   [index: number]: {
@@ -15,6 +16,28 @@ interface Assistant {
   styleUrls: ['./assistants.component.css']
 })
 export class AssistantsComponent implements OnInit {
+
+  newColor = false;
+  toggleColor() {
+    this.newColor = !this.newColor;
+  }
+  stateFlag = false;
+
+  toggleState() {
+    this.stateFlag = !this.stateFlag;
+  }
+
+  styleChange() {
+
+  }
+
+  calculateClasses() {
+    return {
+      'mat-raised-button': true,
+      'btn-extra-class': this.stateFlag
+    };
+  }
+
   assistants: Assistant = [
     { id: '1', name: 'Clark', isAvailable: true },
     { id: '2', name: 'Kim', isAvailable: false },
