@@ -5,12 +5,16 @@ import { ProductService } from './product.service';
   selector: 'products',
   template: `
     <h2>Products</h2>
-    <ul>
-      <li *ngFor="let product of products">
-        {{product}}
-      </li>
-    </ul>
-    `
+    <div *ngIf="products.length == 0; else loading">
+      No products to display
+    </div>
+    <ng-template #loading>
+      <div *ngFor="let product of products">
+        <product [data]="product"></product>
+      </div>
+    </ng-template>
+  `,
+  providers: [ProductService]
 })
 export class ProductsComponent{
   products;
